dx: extract Hosting from Bridge.js to component
Per discussion with ~fallyn-balfus, the Hosting project is on hold.
So, removing it from Home for now by extracting it to a new component
